fixed issue where community imagw was deleted before update

+++ convex/community.ts
@@ -202,9 +202,10 @@ export const mutateCommunityImage = mutation({
 
     if (!isCommunityCreator)
       throw new Error("You are not allowed to update thi community's image");
-    // delete the old communities photo from file storage
-    await ctx.storage.delete(communityRef.communityImage as string);
     // update the communities image
     await ctx.db.patch(communityRef._id, { communityImage: args.storageId });
+    // delete the old communities photo from file storage
+    if (communityRef?.communityImage === "") return;
+    await ctx.storage.delete(communityRef.communityImage as string);
   },
 });
